# apps/backend/core/core-api/src/main/kotlin/swm/calender/core/api/controller/v1/team/request/TeamMemberRoleChangeRequest.kt
package swm.calender.core.api.controller.v1.team.request

import jakarta.validation.constraints.NotNull
import swm.calender.core.enums.TeamMemberRole

data class TeamMemberRoleChangeRequest(
    @field:NotNull
    val role: TeamMemberRole?,
)