# specs/001-swm-teams-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SWM Teams MVP API
  version: 0.1.0
  description: REST contract for the SWM Teams MVP backend.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /teams:
    post:
      tags: [Teams]
      summary: Create team
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreateRequest'
      responses:
        '200':
          description: Team created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponseEnvelope'
  /teams/join:
    post:
      tags: [Teams]
      summary: Join team by invite code
      operationId: joinTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamJoinRequest'
      responses:
        '200':
          description: Team joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponseEnvelope'
  /teams/{teamId}/members:
    get:
      tags: [Teams]
      summary: List team members
      operationId: listTeamMembers
      parameters:
        - $ref: '#/components/parameters/teamId'
      responses:
        '200':
          description: Team members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberListEnvelope'
  /teams/{teamId}/members/{memberId}:
    delete:
      tags: [Teams]
      summary: Remove team member
      operationId: removeTeamMember
      parameters:
        - $ref: '#/components/parameters/teamId'
        - $ref: '#/components/parameters/memberId'
      responses:
        '200':
          description: Team member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberResponseEnvelope'
  /teams/{teamId}/members/{memberId}/role:
    patch:
      tags: [Teams]
      summary: Change member role
      operationId: changeMemberRole
      parameters:
        - $ref: '#/components/parameters/teamId'
        - $ref: '#/components/parameters/memberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberRoleChangeRequest'
      responses:
        '200':
          description: Role changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberResponseEnvelope'
  /teams/{teamId}/sub-services/{subService}:
    patch:
      tags: [Teams]
      summary: Enable or disable a team sub-service
      operationId: updateSubServiceActivation
      parameters:
        - $ref: '#/components/parameters/teamId'
        - name: subService
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/SubService'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubServiceActivationRequest'
      responses:
        '200':
          description: Activation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubServiceActivationEnvelope'
  /calendar/mentoring-schedules:bulk-push:
    post:
      tags: [Calendar]
      summary: Bulk push mentoring schedules to team calendar
      operationId: bulkPushMentoringSchedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentoringScheduleBulkPushRequest'
      responses:
        '200':
          description: Bulk push result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentoringScheduleBulkPushEnvelope'
  /calendar/when2meet-link:
    put:
      tags: [Calendar]
      summary: Register or replace When2meet link
      operationId: putWhen2meetLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/When2meetLinkRequest'
      responses:
        '200':
          description: Link registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/When2meetLinkEnvelope'
  /calendar/availability:
    get:
      tags: [Calendar]
      summary: Get unified team availability
      operationId: getUnifiedAvailability
      parameters:
        - name: startsAt
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endsAt
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Unified availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedAvailabilityEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    teamId:
      name: teamId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    memberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ApiResult:
      type: string
      enum: [SUCCESS, ERROR]
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
    TeamRole:
      type: string
      enum: [OWNER, MEMBER]
    SubService:
      type: string
      enum: [calendar, match]
...
    TeamCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string
          maxLength: 500
    TeamJoinRequest:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
          minLength: 1
          maxLength: 64
    TeamMemberRoleChangeRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/TeamRole'
    SubServiceActivationRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    TeamResponse:
      type: object
      required: [teamId, name, inviteCode, calendarEnabled, matchEnabled]
      properties:
        teamId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        inviteCode:
          type: string
        calendarEnabled:
          type: boolean
        matchEnabled:
          type: boolean
    TeamMemberResponse:
      type: object
      required: [memberId, userId, name, email, role]
      properties:
        memberId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/TeamRole'
    SubServiceActivationResponse:
      type: object
      required: [teamId, calendarEnabled, matchEnabled]
      properties:
        teamId:
          type: integer
          format: int64
        calendarEnabled:
          type: boolean
        matchEnabled:
          type: boolean
    MentoringScheduleBulkPushRequest:
      type: object
      required: [schedules]
      properties:
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/MentoringScheduleRequestItem'
    MentoringScheduleRequestItem:
      type: object
      required: [externalSourceId, title, startsAt, endsAt]
      properties:
        externalSourceId:
          type: string
        title:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    MentoringScheduleBulkPushResponse:
      type: object
      required: [createdCount, skippedDuplicateCount]
      properties:
        createdCount:
          type: integer
          minimum: 0
        skippedDuplicateCount:
          type: integer
          minimum: 0
    When2meetLinkRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
    When2meetLinkResponse:
      type: object
      required: [url, status]
      properties:
        url:
          type: string
        status:
          type: string
          enum: [PENDING, PARSED, FAILED]
        failureReason:
          type: string
          nullable: true
    AvailabilitySlotResponse:
      type: object
      required: [startsAt, endsAt, availableMemberCount, busyMemberCount]
      properties:
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        availableMemberCount:
          type: integer
          minimum: 0
        busyMemberCount:
          type: integer
          minimum: 0
    UnifiedAvailabilityResponse:
      type: object
      required: [slots, generatedAt]
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilitySlotResponse'
        generatedAt:
          type: string
          format: date-time
...
    TeamResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TeamResponse'
    TeamMemberResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TeamMemberResponse'
    TeamMemberListEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/TeamMemberResponse'
    SubServiceActivationEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SubServiceActivationResponse'
    MentoringScheduleBulkPushEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MentoringScheduleBulkPushResponse'
    When2meetLinkEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/When2meetLinkResponse'
    UnifiedAvailabilityEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UnifiedAvailabilityResponse'
...
    EnvelopeBase:
      type: object
      required: [result, error]
      properties:
        result:
          $ref: '#/components/schemas/ApiResult'
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorBody'
            - type: 'null'